bug en preprocessing functions y mejora en página usser

+++ backend/saved_ml_models/preprocessing_functions.py
@@ -75,7 +75,7 @@ class BestModelProductQuantity(BaseModel):
     age: int 
     education: str
     marital_status: str 
-    income: int 
+    income: float 
     kidhome: int 
     teenhome: int 
     year_customer_entered: str
